Add leader companion process
An attempt to ensure there is a companion process that co-exists with
the leader.

It works.

However, its main use case is for at-least-once quorum queue
dead-lettering where queue policies can change dynamically.
So far, in Ra, a server needs to be restarted to change mutable config.
What we need instead is another effect that gets executed on leader and
all follower Ra nodes that changes tha Ra server config (not the State
machine config) dynamically. It's feasible to implement that. The
challenge is that such a config change should not be best-effort.
It must be guaranteed to be executed on all Ra nodes eventually.

Instead of going down that road, for now, we make use of the state enter
callbacks in rabbit_fifo to start and terminate the companion child
process.